Added Notification, Authorization for Admin and Friendship.

+++ TangoApp/TangoApp/Controllers/PostsController.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.AspNet.Identity;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
@@ -11,35 +12,49 @@ namespace TangoApp.Controllers
     {
         private ApplicationDbContext db = new ApplicationDbContext();
         // GET: Posts
+        [Authorize(Roles ="User,Editor,Admin")]
         public ActionResult Index()
         {
-            var posts = db.Posts;
+            var posts = db.Posts.Include("User");
             ViewBag.Posts = posts;
+            if (TempData.ContainsKey("message"))
+            {
+                ViewBag.Message = TempData["message"];
+            }
             return View();
         }
-
+        [Authorize(Roles = "User,Editor,Admin")]
         public ActionResult Show(int id)
         {
             Post post = db.Posts.Find(id);
-
+            if (TempData.ContainsKey("message"))
+            {
+                ViewBag.Message = TempData["message"];
+            }
+            return View(post);
+        }
+        [Authorize(Roles = "Editor,Admin")]
+        public ActionResult New()
+        {
+            Post post = new Post();
+            post.UserId = User.Identity.GetUserId();
             return View(post);
         }
 
-        public ActionResult New()
-        {
-            return View();
-        }
-
         [HttpPost]
+        [Authorize(Roles = "Editor,Admin")]
         public ActionResult New(Post post)
         {
+           
             try
             {
                 if (ModelState.IsValid)
                 {
                     post.Date = DateTime.Now;
+                    post.UserId = User.Identity.GetUserId();
                     db.Posts.Add(post);
                     db.SaveChanges();
+                    TempData["message"] = "Postarea a fost adaugata!";
                     return RedirectToAction("Index");
                 }
                 else
@@ -53,12 +68,75 @@ namespace TangoApp.Controllers
                 return View();
             }
         }
+        [HttpPost]
+        public ActionResult NewComment(int id, Comment com)
+        {
+
+
+            try
+            {
+                if (ModelState.IsValid)
+                {
+                    com.Date = DateTime.Now;
+                    com.UserId = User.Identity.GetUserId();
+                    db.Comments.Add(com);
+                    db.SaveChanges();
+                    TempData["message"] = "Comentariul a fost adaugat!";
+                    //adaugam notificarea corespunzatoare
+                    try
+                    {
+                        Notification notification = new Notification();
+                        notification.UserSendId = User.Identity.GetUserId();
+                        var post = db.Posts.Find(id);
+                        notification.UserReceiveId = post.UserId;
+                        notification.PostId = id;
+                        notification.CommentId = com.CommentId;
+                        notification.Time = DateTime.Now;
+                        notification.Seen = false;
+                        db.Notifications.Add(notification);
+                        db.SaveChanges();
+                    }
+                    catch(Exception e)
+                    {
+                        TempData["message"] = " Eroare la adaugarea notificarii" + e;
+
+                    }
+
+
+                    return Redirect("/Posts/Show/" + com.PostId);
+                }
+                else
+                {
+                    Post post = db.Posts.Find(id);
+                    ViewBag.Comment = com;
+                    return View("Show", post);
+
+                }
+
+            }
+            catch (Exception e)
+            {
+
+                Post post = db.Posts.Find(id);
+                return View("Show", post);
+            }
+
+
+        }
+        [Authorize(Roles = "Editor,Admin")]
         public ActionResult Edit(int id)
         {
             var post = db.Posts.Find(id);
-            return View(post);
+            if(post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+                return View(post);
+            else
+            {
+                TempData["message"] = "Nu aveti permisiunea de a edita aceasta postare!";
+                return RedirectToAction("Index");
+            }
         }
         [HttpPut]
+        [Authorize(Roles = "Editor,Admin")]
         public ActionResult Edit(int id, Post requestpost)
         {
 
@@ -68,20 +146,28 @@ namespace TangoApp.Controllers
                 if(ModelState.IsValid)
                 {
                     var post = db.Posts.Find(id);
-                    if(TryValidateModel(post))
+                    if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
                     {
+                        if (TryValidateModel(post))
+                        {
 
-                        post.Text = requestpost.Text;
-                        post.LastEditDate = DateTime.Now;
-                        db.SaveChanges();
-                        TempData["message"] = "Postarea a fost editata!";
-                        return Redirect("/Show/" + id);
+                            post.Text = requestpost.Text;
+                            post.LastEditDate = DateTime.Now;
+                            db.SaveChanges();
+                            TempData["message"] = "Postarea a fost editata!";
+                            return Redirect("/Posts/Show/" + id);
 
+                        }
+                        else
+                        {
+                            return View(requestpost);
+
+                        }
                     }
                     else
                     {
-                        return View(requestpost);
-
+                        TempData["message"] = "Nu aveti permisiunea de a edita aceasta postare!";
+                        return RedirectToAction("Index");
                     }
 
 
@@ -100,12 +186,22 @@ namespace TangoApp.Controllers
             }
         }
         [HttpDelete]
+        [Authorize(Roles = "Editor,Admin")]
         public ActionResult Delete(int id)
         {
             var post = db.Posts.Find(id);
-            db.Posts.Remove(post);
-            TempData["message"] = "Postarea a fost stearsa!";
-            return RedirectToAction("Index");
+            if (post.UserId == User.Identity.GetUserId() || User.IsInRole("Admin"))
+            {
+                db.Posts.Remove(post);
+                db.SaveChanges();
+                TempData["message"] = "Postarea a fost stearsa!";
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                TempData["message"] = "Nu aveti permisiunea de a sterge aceasta postare!";
+                return RedirectToAction("Index");
+            }
 
         }
